refactor: move a lot of files for domain code
