+++ web/src/lib/types/widget.ts
@@ -1,28 +1,10 @@
+import { type WidgetType } from "~/server/domain/config/widgetType";
 import { type AdjustedWidgetLayout } from "~/server/domain/layout/adjustedWidgetLayout";
-import type { ScreenSize } from "./types";
-
-// general
-export type WidgetType = "calendar" | "search" | "links" | "time";
-
-export type Positioning = {
-  x: number;
-  y: number;
-  w: number;
-  h: number;
-};
-
-export type ScreenSizePositioning = Record<ScreenSize, Positioning>;
-
-export type PartialScreenSizePositioning = {
-  [K in ScreenSize]?: Positioning;
-};
-export type Layout = Positioning | PartialScreenSizePositioning;
+import { type RGLayout } from "~/server/domain/layout/layout";
 
 export type WidgetProps = {
   widget: AdjustedWidgetLayout;
 };
-
-export type RGLayout = ReactGridLayout.Layouts;
 
 export type DisplayedWidgets = {
   rgLayout: RGLayout;
