+++ web/src/lib/ui/commands/commandManager.ts
@@ -1,7 +1,7 @@
 import type ReactGridLayout from "react-grid-layout";
-import { type ScreenSize } from "~/lib/types/types";
-import { type RGLayout } from "~/lib/types/widget";
 import { type AdjustedWidgetLayout } from "~/server/domain/layout/adjustedWidgetLayout";
+import { type RGLayout } from "~/server/domain/layout/layout";
+import { type ScreenSize } from "~/server/domain/other/screenSize";
 import AbortEditCommand from "./abortEditCommand";
 import ChangeWidgetCommand from "./changeWidgetCommand";
 import { type Command } from "./command";
