+++ web/src/lib/ui/hooks.ts
@@ -4,11 +4,13 @@ import { useContext, useEffect, useState } from "react";
 import filterWidgetLayoutByRgl from "~/application/layout/filterWidgetLayoutByLayout.service";
 import getHidingClasses from "~/application/layout/getHidingClasses.service";
 import { type AdjustedWidgetLayout } from "~/server/domain/layout/adjustedWidgetLayout";
+import { type RGLayout } from "~/server/domain/layout/layout";
+import { type ScreenSize } from "~/server/domain/other/screenSize";
 import { api } from "../api/api";
 import { BREAKPOINTS } from "../basic/const";
 import Log from "../log/log";
-import type { ScreenSize, ToastType } from "../types/types";
-import { type DisplayedWidgets, type RGLayout } from "../types/widget";
+import type { ToastType } from "../types/types";
+import { type DisplayedWidgets } from "../types/widget";
 import { CommandContext } from "./context/command";
 import { useBoundStore } from "./state";
 import { toastTextAtom, toastTypeAtom } from "./state/atoms";
