+++ web/src/server/domain/config/widgetConfig.ts
@@ -1,8 +1,13 @@
 import { z } from "zod";
 import { isObject, isString } from "~/lib/guards/base";
-import { isFittingDataPaylod, isWidgetType } from "~/lib/guards/widgets";
-import { widgetTypeSchema } from "~/lib/types/schema";
-import type { WidgetType } from "~/lib/types/widget";
+import {
+  WidgetTypeHelper,
+  type WidgetType,
+} from "~/server/domain/config/widgetType";
+import { isCalendarWidgetConfig } from "~/server/widgets/calendar/guards";
+import { isLinkWidgetConfig } from "~/server/widgets/links/guards";
+import { isSearchWidgetConfig } from "~/server/widgets/search/guards";
+import { isTimeWidgetConfig } from "~/server/widgets/time/guards";
 import { calendarWidgetConfigSchema } from "../../widgets/calendar/schema";
 import type { CalendarWidgetConfig } from "../../widgets/calendar/types";
 import { linkWidgetConfigSchema } from "../../widgets/links/schema";
@@ -26,7 +31,7 @@ export const WidgetConfigHelper = {
   getSchema() {
     const widgetConfigSchema = z.object({
       id: z.string(),
-      type: widgetTypeSchema,
+      type: WidgetTypeHelper.getSchema(),
       data: linkWidgetConfigSchema
         .or(searchWidgetConfigSchema)
         .or(timeWidgetConfigSchema)
@@ -56,13 +61,36 @@ export const WidgetConfigHelper = {
     if (!isString(input.id)) {
       return false;
     }
-    if (!isWidgetType(input.type)) {
+    if (!WidgetTypeHelper.validate(input.type)) {
       return false;
     }
-    if (!isFittingDataPaylod(input.data, input.type)) {
+    if (!this.isFittingDataPaylod(input.data, input.type)) {
       return false;
     }
 
     return true;
+  },
+  /**
+   * Checks whether data is of type  @type {WidgetConfig["data"]}
+   * @param {unknown} data Unkown type to be checked
+   * @param {WidgetType} type Unkown type to be checked
+   * @returns {boolean} Whether data is of type @type {WidgetConfig["data"]}
+   */
+  isFittingDataPaylod(
+    data: unknown,
+    type: WidgetType,
+  ): data is WidgetConfig["data"] {
+    switch (type) {
+      case "calendar":
+        return isCalendarWidgetConfig(data);
+      case "search":
+        return isSearchWidgetConfig(data);
+      case "links":
+        return isLinkWidgetConfig(data);
+      case "time":
+        return isTimeWidgetConfig(data);
+      default:
+        return false;
+    }
   },
 };
